회원가입 로그인 설정

+++ src/main/java/com/prj1/service/BoardService.java
@@ -1,8 +1,10 @@
 package com.prj1.service;
 
 import com.prj1.domain.Board;
+import com.prj1.domain.CustomerUser;
 import com.prj1.mapper.BoardMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,8 +16,11 @@ import java.util.List;
 public class BoardService {
     private final BoardMapper mapper;
 
-    public void addBoard(Board board) {
-        mapper.insert(board);
+    public void addBoard(Board board, Authentication authentication) {
+        if (authentication.getPrincipal() instanceof CustomerUser user) {
+            board.setMemberId(user.getMember().getId());
+            mapper.insert(board);
+        }
     }
 
     public Board selectbyIdBoard(Integer id) {
@@ -26,11 +31,17 @@ public class BoardService {
         return mapper.selectAll();
     }
 
-    public void updateBoard(Board board) {
-        mapper.update(board);
+    public void updateBoard(Board board, Authentication authentication) {
+        if (authentication.getPrincipal() instanceof CustomerUser user) {
+            board.setMemberId(user.getMember().getId());
+            mapper.update(board);
+        }
     }
 
-    public void delete(Board board) {
-        mapper.delete(board);
+    public void delete(Board board, Authentication authentication) {
+        if (authentication.getPrincipal() instanceof CustomerUser user) {
+            board.setMemberId(user.getMember().getId());
+            mapper.delete(board);
+        }
     }
 }
